refactor: modularize reader module into separate components
Split the monolithic reader.rs (737 lines) into focused modules:
- token.rs: Token, OwnedToken, SourceLoc, and TokenWithLoc types
- lexer.rs: Lexer for tokenization and low-level parsing
- parser.rs: Reader for high-level parsing from token stream to Value
- mod.rs: Public API entry point (read_str) and test suite

Benefits:
- Clear separation of concerns between tokenization and parsing
- Lexer is now independently testable and reusable
- Better code organization makes future enhancements easier
- No functional changes; all tests pass
